FY-146/refactor: shouldHideNavbar 함수 분리
